### 2019-07-11 #1560 @arfon
@LarsHernandez - this PR should fix your paper compilation: https://github.com/HaemAalborg/ccostr/pull/8
