### 2019-07-11 #1560 @csoneson
> 👋 @csoneson - would you be willing to edit this submission for JOSS?

Yes, I can do that!
