### 2019-07-15 #1560 @csoneson
👋 @LarsHernandez - it looks like whedon is not happy with the missing date field in your yaml header, could you add that? I compiled it locally to have a first look and it looks almost all good; could you just add case protection for {SAS} in the Zhao and Wang (2010) reference? Also, the word "naïve" doesn't show up properly in my locally compiled paper - whedon might have the same problem. Once you have pushed a new version to your GitHub repo, you can ask whedon to rebuild the pdf using the command above (@whedon generate pdf).
