### 2019-07-21 1c60fb3
fixed naive
