+++ paper.md
@@ -40,7 +40,7 @@ Previous implementations of these estimators into statistical software have been
 
 
 # Estimators
-The R package ccostr includes four different estimators of the mean cost. The average sample, $AS$, estimator simply averages the total cost per individual, disregarding censoring, giving a downwards biased estimate since costs after censoring are not accounted for. The complete case, $CC$, estimator averages the cost of only fully observed cases, biasing the estimate towards the average cost for individuals with shorter survival, typically downwards biased. These two naÃ¯ve estimators are included as reference for the estimators accounting for censoring. For dealing with censored data we implement the $BT$ and $ZT$ estimators for handling situations with or without cost histories. 
+The R package ccostr includes four different estimators of the mean cost. The average sample, $AS$, estimator simply averages the total cost per individual, disregarding censoring, giving a downwards biased estimate since costs after censoring are not accounted for. The complete case, $CC$, estimator averages the cost of only fully observed cases, biasing the estimate towards the average cost for individuals with shorter survival, typically downwards biased. These two naive estimators are included as reference for the estimators accounting for censoring. For dealing with censored data we implement the $BT$ and $ZT$ estimators for handling situations with or without cost histories. 
 
 Assume we observe $\left\{(T_i, \Delta_i, [M_i(u), 0 \leq u\leq T_i]), i=1,...,n\right\}$, where $n$ is the number of individuals, $T_i$ is the observation time, $M_i(u)$ the cost until time $u$, and $\Delta_i$ is event indicator for individual $i$, with $\Delta_i = 1$ or $\Delta_i = 0$ for fully observed and censored cases, respectively. Then the estimates are given by:
 
