### 2019-07-21 #1560 @csoneson
Thanks @LarsHernandez, yes, now it renders correctly. For the reviewers - I think you checked the list of JOSS editors. In the first post in this issue there is a link to a list of people who have signed up as potential reviewers. Could you have a look at that one? I will do the same, and we'll find some good reviewers for your submission!
