### 2019-07-24 #1560 @csoneson
👋@BarryDeCicco, @joethorley - would you be interested in reviewing this submission for JOSS?
