### 2019-07-24 #1560 @BarryDeCicco
Yes - this will be my first review.

    On Wednesday, July 24, 2019, 7:06:13 AM EDT, Charlotte Soneson <notifications@github.com> wrote:

👋@BarryDeCicco, @joethorley - would you be interested in reviewing this submission for JOSS?
