### 2019-07-24 #1560 @joethorley
> 👋@BarryDeCicco, @joethorley - would you be interested in reviewing this submission for JOSS?

I'm happy to review.
