### 2019-07-24 #1560 @csoneson
@BarryDeCicco, @joethorley - thanks a lot! I will close this pre-review issue now and open the main review issue, where the review will take place.
