### 2019-07-24 #1593 @whedon
[REVIEW]: ccostr: An R package for estimating mean costs with censored data
**Submitting author:** @LarsHernandez (<a href="http://orcid.org/0000-0002-3715-8528">Lars Nielsen</a>)
**Repository:** <a href="https://github.com/HaemAalborg/ccostr" target ="_blank">https://github.com/HaemAalborg/ccostr</a>
**Version:** v0.1.0
