### 2019-07-24 #1593 @csoneson
@BarryDeCicco, @joethorley - please perform your reviews here (see instructions in the issue above). Don't hesitate to ping me if you have any questions!
