### 2019-07-24 HaemAalborg/ccostr#9 @joethorley
contributors are not credited
Can @LarsHernandez add @LasseHjort and @rfbrondum as ctb or aut as appropriate?

http://r-pkgs.had.co.nz/description.html#author

https://github.com/openjournals/joss-reviews/issues/1593
