### 2019-07-24 HaemAalborg/ccostr#10 @joethorley
tests fail on my computer due to use of `setwd("~/GitHub/ccostr")`
tests fail on my computer due to use of

setwd("~/GitHub/ccostr")
load(file = "tests/testthat/StataDataTest.Rdata")

http://r-pkgs.had.co.nz/data.html

https://github.com/openjournals/joss-reviews/issues/1593
