### 2019-07-24 HaemAalborg/ccostr#11 @joethorley
SAS needs capitalizing in Analysis of observational health care data using sas
Analysis of observational health care data using sas

https://github.com/openjournals/joss-reviews/issues/1593
