### 2019-07-24 HaemAalborg/ccostr#12 @joethorley
Please .Rbuildignore non-standard files or place in appropriate directories
Non-standard files/directories found at top level:
    ‘README.Rmd’ ‘ccostrRef.bib’ ‘fig’ ‘img’ ‘paper.md’

https://github.com/openjournals/joss-reviews/issues/1593
