### 2019-07-24 HaemAalborg/ccostr#14 @joethorley
please include StataDataTest.Rdata in appropriate directory
please include StataDataTest.Rdata in appropriate directory

http://r-pkgs.had.co.nz/data.html

https://github.com/openjournals/joss-reviews/issues/1593
