### 2019-07-24 #1593 @joethorley
@csoneson and @LarsHernandez

I've made a start on reviewing and filed some issues that require addressing.

I would strongly recommend that the authors study the freely available R packages by Hadley Wickham <http://r-pkgs.had.co.nz> and implement the best practices to R package development detailed in this book.

Although the methods are worthy of publication in JOSS the packaging of the code needs improvement.
