### 2019-07-25 #1593 @BarryDeCicco
Please note that this is my first JOSS review, so don't expect perfection.Also, I got hit by a barrage of e-mails, in a very transactional, step-by-step format.

   - Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.  I was not able to find the list of dependencies; the software installed correctly.

   -

   -  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?  I was not able to find/understand that.

   -  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?   I was not able to find them.  Furthermore, in one section there was a reference to a data set 'df_1', which I didn't find out how to create.

Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support   No
 References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

I didn't find any references to external test data sets.

    On Wednesday, July 24, 2019, 4:13:16 PM EDT, whedon <notifications@github.com> wrote:

Submitting author: @LarsHernandez (Lars Nielsen)
Repository: https://github.com/HaemAalborg/ccostr
Version: 0.1.0
Editor: @csoneson
Reviewer: @BarryDeCicco, @joethorley
Archive: Pending

Status

Status badge code:

Reviewers and authors:

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

Reviewer instructions & questions

   - Make sure you're logged in to your GitHub account
   - Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations

 ✨ Please try and complete your review in the next two weeks ✨

Review checklist for @BarryDeCicco

Conflict of interest

   -  As the reviewer I confirm that I have read the JOSS conflict of interest policy and that there are no conflicts of interest for me to review this work.

Code of Conduct

   -  I confirm that I read and will adhere to the JOSS code of conduct.

General checks

   -  Repository: Is the source code for this software available at the repository url?
   -  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
   -  Version: Does the release version given match the GitHub release (0.1.0)?
   -  Authorship: Has the submitting author (@LarsHernandez) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Functionality

   -  Installation: Does installation proceed as outlined in the documentation?
   -  Functionality: Have the functional claims of the software been confirmed?
   -  Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

Documentation

   -  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
   -  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
   -  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
   -  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
   -  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
   -  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Software paper

   -  Authors: Does the paper.md file include a list of authors with their affiliations?
   -  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
   -  References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

Review checklist for @joethorley

Conflict of interest

   -  As the reviewer I confirm that I have read the JOSS conflict of interest policy and that there are no conflicts of interest for me to review this work.

Code of Conduct

   -  I confirm that I read and will adhere to the JOSS code of conduct.

General checks

   -  Repository: Is the source code for this software available at the repository url?
   -  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
   -  Version: Does the release version given match the GitHub release (0.1.0)?
   -  Authorship: Has the submitting author (@LarsHernandez) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Functionality

   -  Installation: Does installation proceed as outlined in the documentation?
   -  Functionality: Have the functional claims of the software been confirmed?
   -  Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

Documentation

   -  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
   -  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
   -  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
   -  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
   -  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
   -  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Software paper

   -  Authors: Does the paper.md file include a list of authors with their affiliations?
   -  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
   -  References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
