### 2019-07-25 #1593 @BarryDeCicco
I have sent some comments; if you didn't get them, please let me know (there are several aliased addresses, so I don't know who is what).
Barry

    On Wednesday, July 24, 2019, 7:23:28 PM EDT, Joe Thorley <notifications@github.com> wrote:

@csoneson and @LarsHernandez

I've made a start on reviewing and filed some issues that require addressing.

I would strongly recommend that the authors study the freely available R packages by Hadley Wickham http://r-pkgs.had.co.nz and implement the best practices to R package development detailed in this book.

Although the methods are worthy of publication in JOSS the packaging of the code needs improvement.
