### 2019-07-25 #1593 @csoneson
@BarryDeCicco Thanks, your comments showed up here. Probably the easiest is to comment directly in this GitHub issue (click on "view it on GitHub" in the bottom of your email), that avoids having to worry about email addresses etc. Also note that there is a checklist in the beginning of this issue, where you should check off the aspects that you find satisfactory in the submission (and either provide comments in this issue or open issues in the submitted repository for the points that you do not find satisfactorily addressed). As the authors update the submission, please keep ticking the boxes as you consider them fulfilled. All boxes need to be ticked by all reviewers in order to eventually accept the paper.
