### 2019-07-29 #1593 @csoneson
ping @BarryDeCicco
> @BarryDeCicco - could you please clarify? Do you no longer wish to be a reviewer for this submission? And/or do you want to be removed from the list of JOSS reviewers for future submissions?
