### 2019-08-01 #1593 @csoneson
@zedtaha - would you be interested in reviewing this submission for JOSS?
