### 2019-08-07 HaemAalborg/ccostr#9 @LarsHernandez
I updated the DESCRIPTION file

![image](https://user-images.githubusercontent.com/16371910/62601957-d9f7c180-b8f2-11e9-8419-9431f504d205.png)
