### 2019-08-07 HaemAalborg/ccostr#12 @LarsHernandez
Thanks for the comment, but Is this a problem for the JOSS submission? It would nice to keep these files there, or do i need two repositories?
