### 2019-08-07 HaemAalborg/ccostr#14 @LarsHernandez
From the link it says:
![image](https://user-images.githubusercontent.com/16371910/62602663-84241900-b8f4-11e9-8c1a-5d99267947bf.png)

the file i include is 401kb, i'm not sure if this is considered small?

![image](https://user-images.githubusercontent.com/16371910/62602815-e67d1980-b8f4-11e9-811a-cd0d8041f589.png)

The stata files (.dta) in the StataTestCcostr folder i will delete, but in the directory i also have two more files, one text file with the stata code i ran, and one with the code to generate the data, should i just remove the whole folder?

![image](https://user-images.githubusercontent.com/16371910/62602838-f0068180-b8f4-11e9-821f-a0411ef0ecbe.png)
