### 2019-08-07 HaemAalborg/ccostr#11 @LarsHernandez
Thanks, it should be fixed now
