### 2019-08-07 3857931
updating tests and SAS reference
