+++ ccostrRef.bib
@@ -127,7 +127,7 @@ year = {2001}
 
 @incollection{Zhao2010,
 author = {Zhao, Honwei and Wang, Hongkun},
-booktitle = {Analysis of Observational Health Care Data Using SAS},
+booktitle = {Analysis of Observational Health Care Data Using {SAS}},
 editor = {Faries, Douglas E. and Obenchain, R. and Haro, Josep M. and Leon, Andrew C.},
 pages = {363--381},
 publisher = {SAS Press},
