### 2019-08-07 HaemAalborg/ccostr#10 @LarsHernandez
I moved the test data to inst/extdata and used:
![image](https://user-images.githubusercontent.com/16371910/62606382-6955a280-b8fc-11e9-8384-90f7e111c8be.png)

I think this works
