### 2019-08-07 HaemAalborg/ccostr#14 @LarsHernandez
I moved the StataDataTest file to fix the relative path issue, still not sure if i should still include the scripts to replicate the data in R and estimation in Stata..
