### 2019-08-07 #1593 @csoneson
@tgerke - would you be interested in reviewing this submission for JOSS?
