### 2019-08-07 #1593 @LarsHernandez
Hey Charlotte, thank you for your effort so far. I just returned from vacation and I've just been fixing some of the Github issues raised by @joethorley, and looked into the comments from @BarryDeCicco, thanks a lot for the issues/comments. This is my first submission/article, so i'm very new in this, but i guess the submission needs two reviewers right? If @tgerke doesn't answer i can try to suggest one, i think the primary focus is R and survival analysis, i can see two with both in the list.
