### 2019-08-07 HaemAalborg/ccostr#12 @joethorley
You can keep them there you just need to make sure R ignores them (without a NOTE) by listing them in .Rbuildignore
