### 2019-08-07 HaemAalborg/ccostr#14 @joethorley
The limit for an R package is 5MB so you are fine. Just put everything in data_raw.

The best way to setup is using

`usethis::use_data_raw()`
