### 2019-08-07 HaemAalborg/ccostr#10 @joethorley
yes this is appropriate
