### 2019-08-07 HaemAalborg/ccostr#14 @joethorley
and then use `usethis::use_data()` in a script in data raw to import the data into the package.
