### 2019-08-07 #1593 @LarsHernandez
Thanks, and that sounds good, also I've just added a standard code of conduct to the repository which should check the **community guidelines** box
