### 2019-08-12 #1593 @tgerke
This looks really interesting! Happy to help. Will aim to tackle the review late this week, if that timeline works for you all.
