### 2019-08-14 6be8ba7
NSE errors fix and Rbuildignore fix
Removed the arguments in the ccmean function call as they were not working, this created NSE errors that is now fixed
