+++ NAMESPACE
@@ -15,6 +15,7 @@ importFrom(Rdpack,reprompt)
 importFrom(data.table,":=")
 importFrom(data.table,as.data.table)
 importFrom(data.table,setDTthreads)
+importFrom(rlang,.data)
 importFrom(stats,rexp)
 importFrom(stats,runif)
 importFrom(stats,var)
