### 2019-08-14 HaemAalborg/ccostr#12 @LarsHernandez
Finally got it fixed, i learned a lesson never to put spaces in Rbuildignore :S
