+++ .travis.yml
@@ -1,0 +1,15 @@
+language: r
+
+sudo: required
+
+env: _R_CHECK_CRAN_INCOMING_=FALSE
+
+r_packages:
+  - covr
+  - devtools
+
+r_github_packages:
+  - HaemAalborg/ccostr
+
+after_success:
+  - Rscript -e 'library(covr);codecov()'
