### 2019-08-14 4c5231b
badges and raw data (not finished)
