+++ README.Rmd
@@ -15,6 +15,8 @@ library(knitr)
 
 # ccostr
 [![status](http://joss.theoj.org/papers/0b9e631729dd9fadff3d6875e84ed954/status.svg)](http://joss.theoj.org/papers/0b9e631729dd9fadff3d6875e84ed954)
+[![Build Status](https://travis-ci.org/LarsHernandez/ccostr.svg?branch=master)](https://travis-ci.org/LarsHernandez/ccostr)
+[![codecov](https://codecov.io/gh/LarsHernandez/ccostr/branch/master/graph/badge.svg)](https://codecov.io/gh/LarsHernandez/ccostr)
 
 ccostr is an R package to calculate estimates of mean total cost in censored cost data, i.e. in situations where data are not fully observed within the study period.
 
