### 2019-08-22 d989723
fixes to readme and fix of paper "similar" -> "identical"
