### 2019-08-22 f4ed46d
added data to test in SAS and removed some old files
