### 2019-08-22 HaemAalborg/ccostr#14 @LarsHernandez
I've tried to use the usethis approach, but i've had some problems getting it to work. I've moved some things around and i think the datastructure is ok, and it works, so i don't think i'll look more into it for now, but thanks for the suggestion :)
