### 2019-08-22 #1593 @LarsHernandez
Thanks @joethorley and @tgerke for all the comments/issues, I've tried to implement all of the things you have raised. I can see both of you have not checked the Functionality/Example usage boxes, is there anything i/we can do to get that checked? Is the package lacking anything aside from the examples included in the two vignettes?
