### 2019-08-22 #1593 @tgerke
Looking great @LarsHernandez! I just opened two issues regarding the examples in the vignettes, and once those minor points are resolved, it looks all set to me.
