### 2019-08-22 HaemAalborg/ccostr#25 @joethorley
Can you document each of the columns for hcost in the hcost help file.
For an example see

?dplyr::starwars

and for more information

http://r-pkgs.had.co.nz/data.html#documenting-data

openjournals/joss-reviews#1593
