### 2019-08-28 HaemAalborg/ccostr#25 @LarsHernandez
Thanks, I've added it now
