### 2019-08-28 4f34dbf
unhide code chunk
