### 2019-08-28 cfb81d7
Update hcost.Rd
