+++ inst/extdata/SAS/uesurv.txt
@@ -1,0 +1,144 @@
+1,1,7.40089661674574
+1,1,7.40089661674574
+1,1,7.40089661674574
+1,1,7.40089661674574
+1,1,7.40089661674574
+1,1,7.40089661674574
+1,1,7.40089661674574
+1,1,7.40089661674574
+2,1,0.201669754460454
+3,0,3.18330507817052
+3,0,3.18330507817052
+3,0,3.18330507817052
+3,0,3.18330507817052
+4,1,1.11760509433225
+4,1,1.11760509433225
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+5,1,9.25619227578864
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+6,1,9.05117775779217
+7,1,3.97897896822542
+7,1,3.97897896822542
+7,1,3.97897896822542
+7,1,3.97897896822542
+8,0,7.93680119514465
+8,0,7.93680119514465
+8,0,7.93680119514465
+8,0,7.93680119514465
+8,0,7.93680119514465
+8,0,7.93680119514465
+8,0,7.93680119514465
+8,0,7.93680119514465
+9,1,6.63777928799391
+9,1,6.63777928799391
+9,1,6.63777928799391
+9,1,6.63777928799391
+9,1,6.63777928799391
+9,1,6.63777928799391
+9,1,6.63777928799391
+10,0,1.00069240480661
+10,0,1.00069240480661
+11,1,0.0625065318308771
+12,0,3.81207100301981
+12,0,3.81207100301981
+12,0,3.81207100301981
+12,0,3.81207100301981
+13,1,1.92338148131967
+13,1,1.92338148131967
+14,1,4.58195017185062
+14,1,4.58195017185062
+14,1,4.58195017185062
+14,1,4.58195017185062
+14,1,4.58195017185062
+15,1,2.42783789290115
+15,1,2.42783789290115
+15,1,2.42783789290115
+16,0,1.53152950853109
+16,0,1.53152950853109
+17,1,4.15642084786668
+17,1,4.15642084786668
+17,1,4.15642084786668
+17,1,4.15642084786668
+17,1,4.15642084786668
+18,1,4.24837351310998
+18,1,4.24837351310998
+18,1,4.24837351310998
+18,1,4.24837351310998
+18,1,4.24837351310998
+19,0,6.44917485117912
+19,0,6.44917485117912
+19,0,6.44917485117912
+19,0,6.44917485117912
+19,0,6.44917485117912
+19,0,6.44917485117912
+19,0,6.44917485117912
+20,1,5.00874539138749
+20,1,5.00874539138749
+20,1,5.00874539138749
+20,1,5.00874539138749
+20,1,5.00874539138749
+20,1,5.00874539138749
+21,1,4.73466362571344
+21,1,4.73466362571344
+21,1,4.73466362571344
+21,1,4.73466362571344
+21,1,4.73466362571344
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+22,1,8.36495839525014
+23,1,2.76575442403555
[…]
