+++ inst/extdata/SAS/eucost.txt
@@ -1,0 +1,112 @@
+1,0,0.728083211928606,38045.1500005873
+2,0,0.687552923461987,33769.1535304077
+3,0,1,8095.09181026369
+3,1,2,1900.66415788606
+3,2,2.13118040468544,257.292452043288
+4,0,0.0247804932296276,23931.2290589075
+5,0,1,9934.9690085277
+5,1,1.29415946182546,24939.7335689415
+6,0,1,9875.07353238761
+6,1,2,2645.25546915829
+6,2,3,2878.87572105974
+6,3,4,2689.12644162774
+6,4,5,2879.68340208754
+6,5,6,2838.17467372864
+6,6,7,2615.85155175999
+6,7,7.57430640085756,12767.6140809326
+7,0,0.887904709269568,26504.7320789866
+8,0,1,9978.86479757726
+8,1,2,1210.14700382948
+8,2,3,1279.17567407712
+8,3,4,1344.65874582529
+8,4,4.30898400945496,14053.4191860691
+9,0,1,12185.42876672
+9,1,2,2662.312005274
+9,2,2.34005223121494,12277.8662781551
+10,0,1,9859.79946292937
+10,1,1.39209931623191,15175.3552436607
+11,0,1,10155.7874072343
+11,1,1.37439198252128,22642.4154616879
+12,0,0.234880931997653,35412.4235001543
+13,0,1,15364.9668042548
+13,1,2,2757.72231258452
+13,2,2.3362348228693,956.609839930922
+14,0,1,14399.8331558891
+14,1,2,1899.01765352115
+14,2,3,1830.09613184258
+14,3,4,1792.52110598609
+14,4,4.77627390167058,14305.5697145746
+15,0,1,14683.4148414433
+15,1,2,1277.24892599508
+15,2,3,1361.73465652391
+15,3,4,1306.332034152
+15,4,5,1282.55567885935
+15,5,5.37470363055296,19205.3594937647
+16,0,0.705418313038776,39998.6459585116
+17,0,0.562229102477431,27325.9263367976
+18,0,1,15624.462999776
+18,1,2,2285.17592921853
+18,2,3,2436.83639857918
+18,3,4,2378.42600885779
+18,4,5,2364.02033613995
+18,5,5.89590416637189,29338.7221851494
+19,0,1,15997.7889643051
+19,1,2,2208.16507283598
+19,2,3,2483.4054986015
+19,3,4,2313.05784899741
+19,4,4.65834238168724,18938.794247362
+20,0,1,8531.85475021601
+20,1,2,2265.42715076357
+20,2,3,2224.96509468183
+20,3,4,2179.3475728482
+20,4,5,2151.41470525414
+20,5,6,2149.23675451428
+20,6,7,2325.34268433228
+20,7,7.35741068143398,820.577690233031
+21,0,1,8177.04087030143
+21,1,1.86737825721502,30684.1097437811
+22,0,1,7709.17220497504
+22,1,2,1141.30734922364
+22,2,3,1158.22648443282
+22,3,3.88098844792694,1012.63967762764
+23,0,1,9814.71615238115
+23,1,2,2388.61801223829
+23,2,3,2388.62498486415
+23,3,4,2435.06903909147
+23,4,4.06967815943062,163.198800472965
+24,0,1,7465.80690117553
+24,1,2,1823.44858488068
+24,2,3,1623.59045036137
+24,3,4,1874.90563057363
+24,4,5,1939.89520594478
+24,5,6,1951.5789244324
+24,6,6.91365906171847,25440.3739508337
+25,0,1,8800.37958351895
+25,1,2,1711.18640806526
+25,2,3,1459.0377131477
[…]

[…]
