+++ inst/extdata/SAS/uusurv.txt
@@ -1,140 +1,30 @@
 1,1,5.80294366693124
-1,1,5.80294366693124
-1,1,5.80294366693124
-1,1,5.80294366693124
-1,1,5.80294366693124
-1,1,5.80294366693124
 2,1,2.29380094911903
-2,1,2.29380094911903
-2,1,2.29380094911903
-3,1,1.72508574323729
 3,1,1.72508574323729
 4,1,0.373387788422406
 5,1,2.62384495465085
-5,1,2.62384495465085
-5,1,2.62384495465085
-6,1,2.709118751809
-6,1,2.709118751809
 6,1,2.709118751809
 7,1,4.65562499128282
-7,1,4.65562499128282
-7,1,4.65562499128282
-7,1,4.65562499128282
-7,1,4.65562499128282
-8,1,1.52933206409216
 8,1,1.52933206409216
 9,1,0.65952547127381
 10,0,4.54914699774235
-10,0,4.54914699774235
-10,0,4.54914699774235
-10,0,4.54914699774235
-10,0,4.54914699774235
-11,1,3.62280361587182
-11,1,3.62280361587182
-11,1,3.62280361587182
 11,1,3.62280361587182
 12,1,6.59828545292839
-12,1,6.59828545292839
-12,1,6.59828545292839
-12,1,6.59828545292839
-12,1,6.59828545292839
-12,1,6.59828545292839
-12,1,6.59828545292839
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
-13,1,9.65393338352442
 13,1,9.65393338352442
 14,1,2.90861250134185
-14,1,2.90861250134185
-14,1,2.90861250134185
 15,1,4.71093143569306
-15,1,4.71093143569306
-15,1,4.71093143569306
-15,1,4.71093143569306
-15,1,4.71093143569306
-16,0,6.28134100232273
-16,0,6.28134100232273
-16,0,6.28134100232273
-16,0,6.28134100232273
-16,0,6.28134100232273
-16,0,6.28134100232273
 16,0,6.28134100232273
 17,1,2.52283470937982
-17,1,2.52283470937982
-17,1,2.52283470937982
-18,0,2.18664003536105
-18,0,2.18664003536105
 18,0,2.18664003536105
 19,1,6.29904830129817
-19,1,6.29904830129817
-19,1,6.29904830129817
-19,1,6.29904830129817
-19,1,6.29904830129817
-19,1,6.29904830129817
-19,1,6.29904830129817
-20,1,7.99393378430977
-20,1,7.99393378430977
-20,1,7.99393378430977
-20,1,7.99393378430977
-20,1,7.99393378430977
-20,1,7.99393378430977
-20,1,7.99393378430977
 20,1,7.99393378430977
 21,1,4.23949517076835
-21,1,4.23949517076835
-21,1,4.23949517076835
-21,1,4.23949517076835
-21,1,4.23949517076835
-22,0,3.36937790270895
-22,0,3.36937790270895
-22,0,3.36937790270895
 22,0,3.36937790270895
 23,1,2.43530539330095
-23,1,2.43530539330095
-23,1,2.43530539330095
 24,1,5.98452027188614
-24,1,5.98452027188614
-24,1,5.98452027188614
-24,1,5.98452027188614
-24,1,5.98452027188614
-24,1,5.98452027188614
-25,0,7.84328962676227
-25,0,7.84328962676227
-25,0,7.84328962676227
-25,0,7.84328962676227
-25,0,7.84328962676227
[…]
