### 2019-08-29 1a6f917
Update test-SAS_similarity.R
