+++ tests/testthat/test-4_SAS_similarity.R
@@ -63,6 +63,7 @@ test_that("mean estimates sim uu1", {
 })
 test_that("variance estimates sim uu1", {
   expect_equal(round(est.uu1$Estimates$BT[3],2), 2036.03)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.uu1$Estimates$ZT[3],2), 2314.78)
 })
 test_that("mean estimates sim uu2", {
@@ -71,6 +72,7 @@ test_that("mean estimates sim uu2", {
 })
 test_that("variance estimates sim uu2", {
   expect_equal(round(est.uu2$Estimates$BT[3],2), 1438.65)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.uu2$Estimates$ZT[3],2), 1464.40)
 })
 
@@ -86,6 +88,7 @@ test_that("mean estimates sim ee1", {
 })
 test_that("variance estimates sim ee1", {
   expect_equal(round(est.ee1$Estimates$BT[3],2), 3443.89)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.ee1$Estimates$ZT[3],2), 3386.95)
 })
 test_that("mean estimates sim ee2", {
@@ -94,6 +97,7 @@ test_that("mean estimates sim ee2", {
 })
 test_that("variance estimates sim ee2", {
   expect_equal(round(est.ee2$Estimates$BT[3],2), 1801.40)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.ee2$Estimates$ZT[3],2), 1747.04)
 })
 
@@ -110,6 +114,7 @@ test_that("mean estimates sim eu1", {
 })
 test_that("variance estimates sim eu1", {
   expect_equal(round(est.eu1$Estimates$BT[3],2), 1895.46)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.eu1$Estimates$ZT[3],2), 1955.29)
 })
 test_that("mean estimates sim eu2", {
@@ -118,6 +123,7 @@ test_that("mean estimates sim eu2", {
 })
 test_that("variance estimates sim eu2", {
   expect_equal(round(est.eu2$Estimates$BT[3],2), 1808.54)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.eu2$Estimates$ZT[3],2), 1734.09)
 })
 
@@ -133,6 +139,7 @@ test_that("mean estimates sim ue1", {
 })
 test_that("variance estimates sim ue1", {
   expect_equal(round(est.ue1$Estimates$BT[3],2), 1619.20)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.ue1$Estimates$ZT[3],2), 1748.23)
 })
 test_that("mean estimates sim ue2", {
@@ -141,5 +148,6 @@ test_that("mean estimates sim ue2", {
 })
 test_that("variance estimates sim ue2", {
   expect_equal(round(est.ue2$Estimates$BT[3],2), 1520.39)
+  skip('SAS/IML macro implementation of variance on ZT differs')
   expect_equal(round(est.ue2$Estimates$ZT[3],2), 1511.76)
 })
