### 2019-08-29 #1593 @LarsHernandez
update: I've tested the SAS/IML macro that also implements the BT and ZT estimator. The macro was very slow so i created new small datasets to test it, and i get all the same results down to two decimal points, except for the variance of the ZT estimator. It must be implemented a bit different in SAS, but it is very close.
