### 2019-09-03 #1593 @csoneson
@joethorley, @tgerke - thanks a lot for your reviews! I can see that you have both checked all the items in your checklists. @joethorley, based on your comment above I take it that you are happy with the current state of the submission. @tgerke - could you confirm that your final issues have been resolved as well?
