### 2019-09-03 #1593 @csoneson
@LarsHernandez - I have sent a PR with some small modifications to the manuscript - please check and merge (and generate a new proof with `@whedon generate pdf`). Specifically, note that the link to the GitHub repository will be given on the first page of the article, and does not need to be repeated in the text.

Next, please:
- [ ] check that the version information is consistently reported
- [ ] create a tagged release of the software
- [ ] make a Zenodo archive. Note that the archive should have the same title and authors as the article (you can also add authors' ORCID)
- [ ] report back here with the version number and the DOI of the Zenodo archive
