### 2019-09-05 1331714
Delete BTp.R
will be moved to v. 0.1.1
