### 2019-09-05 #1593 @LarsHernandez
- [x] check that the version information is consistently reported:
I've removed some experimentation with a new estimator, BTp, and released it with version 0.1.0.. it passes all tests and the CMD check.. I've created a branch Version-0.1.1 and added the BTp function, so that might be added to the MASTER sometime in the future.

- [x] create a tagged release of the software:
I did the Github release, hope that is what you meant.
