### 2019-09-05 #1593 @LarsHernandez
- [x]  make a Zenodo archive. Note that the archive should have the same title and authors as the article (you can also add authors' ORCID):
[link](https://zenodo.org/record/3386740#.XXC21Cj7RPY)

- [x]  report back here with the version number and the DOI of the Zenodo archive
version: 0.1.0
doi: 10.5281/zenodo.3386740
