### 2019-09-05 #1593 @csoneson
Thanks @LarsHernandez - this looks all good. I'll hand over to the AEiC on duty, who will do the final steps.
@openjournals/joss-eics - this paper is ready to be accepted!
