### 2019-09-05 #1593 @LarsHernandez
Thanks everyone on making my first paper submission so pleasant 😀
